Don't run E2E tests in parallel
Or rather, achieve parallelism by brute-forcing it at the CI job level.

It's possible for some tests to run in parallel at the Go test level,
but I was finding this required careful design and was a bit limiting.

I considered also having each test spin up its own cluster, but I'd
ideally like to make the test framework spinning up a cluster optional
so that you could BYO in future (e.g. to run the suite on EKS or GKE).

Signed-off-by: Nic Cope <nicc@rk0n.org>